removes enough warnings for public release

+++ include/cuda_util.cuh
@@ -163,7 +163,7 @@ void getGridAndBlock(unsigned long numElements, dim3 &grid, unsigned int desired
     }
   }
   else{
-    grid = {numElements,1,1};
+    grid = {(unsigned int) numElements,1,1};
   }
 }
